+++ ydb/core/kqp/common/simple/settings.h
@@ -14,6 +14,9 @@ namespace NKikimr::NKqp {
 struct TKqpQuerySettings {
     bool DocumentApiRestricted = true;
     bool IsInternalCall = false;
+    i32 RuntimeParameterSizeLimit = 0;
+    bool RuntimeParameterSizeLimitSatisfied = false;
+
     NKikimrKqp::EQueryType QueryType = NKikimrKqp::EQueryType::QUERY_TYPE_UNDEFINED;
     Ydb::Query::Syntax Syntax = Ydb::Query::Syntax::SYNTAX_UNSPECIFIED;
 
@@ -25,7 +28,9 @@ struct TKqpQuerySettings {
             DocumentApiRestricted == other.DocumentApiRestricted &&
             IsInternalCall == other.IsInternalCall &&
             QueryType == other.QueryType &&
-            Syntax == other.Syntax;
+            Syntax == other.Syntax &&
+            RuntimeParameterSizeLimit == other.RuntimeParameterSizeLimit &&
+            RuntimeParameterSizeLimitSatisfied == other.RuntimeParameterSizeLimitSatisfied;
     }
 
     bool operator!=(const TKqpQuerySettings& other) {
@@ -38,7 +43,9 @@ struct TKqpQuerySettings {
     bool operator>=(const TKqpQuerySettings&) = delete;
 
     size_t GetHash() const noexcept {
-        auto tuple = std::make_tuple(DocumentApiRestricted, IsInternalCall, QueryType, Syntax);
+        auto tuple = std::make_tuple(
+            DocumentApiRestricted, IsInternalCall, QueryType, Syntax,
+            RuntimeParameterSizeLimitSatisfied);
         return THash<decltype(tuple)>()(tuple);
     }
 
@@ -46,7 +53,10 @@ struct TKqpQuerySettings {
         TStringBuilder result = TStringBuilder() << "{"
             << "DocumentApiRestricted: " << DocumentApiRestricted << ", "
             << "IsInternalCall: " << IsInternalCall << ", "
-            << "QueryType: " << QueryType << "}";
+            << "QueryType: " << QueryType << ", "
+            << "RuntimeParameterSizeLimit: " << RuntimeParameterSizeLimit
+            << "RuntimeParameterSizeLimitSatisfied: " << RuntimeParameterSizeLimitSatisfied
+            << "}";
         return result;
     }
 };