+++ ydb/core/kqp/compile_service/kqp_compile_service.cpp
@@ -359,6 +359,7 @@ private:
         bool defaultSyntaxVersion = TableServiceConfig.GetSqlVersion();
 
         ui64 rangesLimit = TableServiceConfig.GetExtractPredicateRangesLimit();
+        ui64 paramLimitSize = TableServiceConfig.GetExtractPredicateParameterListSizeLimit();
         ui64 idxLookupPointsLimit = TableServiceConfig.GetIdxLookupJoinPointsLimit();
 
         bool allowOlapDataQuery = TableServiceConfig.GetAllowOlapDataQuery();
@@ -400,6 +401,7 @@ private:
         bool enableTempTablesForUser = TableServiceConfig.GetEnableTempTablesForUser();
 
         bool enableSimpleProgramsSinglePartitionOptimization = TableServiceConfig.GetEnableSimpleProgramsSinglePartitionOptimization();
+        bool enableSimpleProgramsSinglePartitionOptimizationBroadPrograms = TableServiceConfig.GetEnableSimpleProgramsSinglePartitionOptimizationBroadPrograms();
 
         ui32 defaultLangVer = TableServiceConfig.GetDefaultLangVer();
 
@@ -429,6 +431,7 @@ private:
             TableServiceConfig.GetEnableDataShardCreateTableAs() != enableDataShardCreateTableAs ||
             TableServiceConfig.GetBlockChannelsMode() != blockChannelsMode ||
             TableServiceConfig.GetExtractPredicateRangesLimit() != rangesLimit ||
+            TableServiceConfig.GetExtractPredicateParameterListSizeLimit() != paramLimitSize ||
             TableServiceConfig.GetResourceManager().GetMkqlHeavyProgramMemoryLimit() != mkqlHeavyLimit ||
             TableServiceConfig.GetIdxLookupJoinPointsLimit() != idxLookupPointsLimit ||
             TableServiceConfig.GetEnableSpillingNodes() != enableSpillingNodes ||
@@ -451,6 +454,7 @@ private:
             TableServiceConfig.GetEnableOlapPushdownProjections() != enableOlapPushdownProjections ||
             TableServiceConfig.GetEnableTempTablesForUser() != enableTempTablesForUser ||
             TableServiceConfig.GetEnableSimpleProgramsSinglePartitionOptimization() != enableSimpleProgramsSinglePartitionOptimization ||
+            TableServiceConfig.GetEnableSimpleProgramsSinglePartitionOptimizationBroadPrograms() != enableSimpleProgramsSinglePartitionOptimizationBroadPrograms ||
             TableServiceConfig.GetDefaultLangVer() != defaultLangVer ||
             TableServiceConfig.GetEnableOlapPushdownAggregate() != enableOlapPushdownAggregate ||
             TableServiceConfig.GetEnableOrderOptimizaionFSM() != enableOrderOptimizaionFSM ||