+++ ydb/library/yql/dq/comp_nodes/dq_hash_join_table.cpp
@@ -1,0 +1,11 @@
+#include "dq_hash_join_table.h"
+
+namespace NKikimr::NMiniKQL::NJoinTable {
+bool NeedToTrackUnusedRightTuples(EJoinKind kind) {
+    return (static_cast<int>(kind)&4) == 4;
+}
+bool NeedToTrackUnusedLeftTuples(EJoinKind kind) {
+    return static_cast<int>(kind)&1 == 1;
+}
+
+}