+++ ydb/library/yql/dq/comp_nodes/dq_scalar_hash_join.cpp
@@ -31,14 +31,38 @@ class TScalarHashJoinState : public TComputationValue<TScalarHashJoinState> {
     IComputationWideFlowNode* ProbeSide() const {
         return LeftFinished_ ? nullptr : LeftFlow_;
     }
-
-    void AppendTuple(NJoinTable::TTuple probe, NJoinTable::TTuple build, std::vector<NUdf::TUnboxedValue>& output){
-        std::copy(probe, probe + LeftColumnTypes_.size(), std::back_inserter(output));
-        for(int index = 0; index < std::ssize(RightColumnTypes_); ++index){
-            if (std::ranges::find(RightKeyColumns_, index) == RightKeyColumns_.end()){
-                output.push_back(build[index]);
+    void AppendTuple(NJoinTable::TTuple probe, NJoinTable::TTuple build, std::vector<NUdf::TUnboxedValue>& output) {
+        MKQL_ENSURE(probe || build,"appending invalid tuple");
+        if (probe) {
+            for (int index = 0; index < std::ssize(LeftKeyColumns_); ++index) {
+                output.push_back(probe[LeftKeyColumns_[index]]);
+            }
+        } else {
+            for (int index = 0; index < std::ssize(RightKeyColumns_); ++index) {
+                output.push_back(build[RightKeyColumns_[index]]);
             }
         }
+
+        for (int index = 0; index < std::ssize(LeftColumnTypes_); ++index) {
+            if (std::ranges::find(LeftKeyColumns_, index) == LeftKeyColumns_.end()) {
+                if (probe) {
+                    output.push_back(probe[index]);                
+                } else {
+                    output.push_back(NYql::NUdf::TUnboxedValuePod{});
+                }
+            }
+        }
+
+        for (int index = 0; index < std::ssize(RightColumnTypes_); ++index) {
+            if (std::ranges::find(RightKeyColumns_, index) == RightKeyColumns_.end()) {
+                if (build) {
+                    output.push_back(build[index]);                
+                } else {
+                    output.push_back(NYql::NUdf::TUnboxedValuePod{});
+                }
+            }
+        }
+
     }
 
 public:
@@ -47,7 +71,7 @@ public:
         IComputationWideFlowNode* leftFlow, IComputationWideFlowNode* rightFlow,
         const std::vector<ui32>& leftKeyColumns, const std::vector<ui32>& rightKeyColumns,
         const std::vector<TType*>& leftColumnTypes, const std::vector<TType*>& rightColumnTypes, [[maybe_unused]] TComputationContext& ctx,
-        NUdf::TLoggerPtr logger, NUdf::TLogComponentId logComponent)
+        NUdf::TLoggerPtr logger, NUdf::TLogComponentId logComponent, EJoinKind joinKind)
     :   TBase(memInfo)
     ,   LeftFlow_(leftFlow)
     ,   RightFlow_(rightFlow)
@@ -58,19 +82,25 @@ public:
     ,   Logger_(logger)
     ,   LogComponent_(logComponent)
     ,   KeyTypes_(KeyTypesFromColumns(leftColumnTypes, leftKeyColumns))
-    ,   Table_(std::ssize(leftColumnTypes), TWideUnboxedEqual{KeyTypes_}, TWideUnboxedHasher{KeyTypes_})
+    ,   JoinKind_(joinKind)
+    ,   Table_(
+        std::ssize(rightColumnTypes)
+        , TWideUnboxedEqual{KeyTypes_}
+        , TWideUnboxedHasher{KeyTypes_}
+        , NJoinTable::NeedToTrackUnusedRightTuples(joinKind))
     ,   Values_(rightColumnTypes.size())
     ,   Pointers_()
     ,   Output_()
     {
         MKQL_ENSURE(RightColumnTypes_.size() == LeftColumnTypes_.size(), "unimplemented");
+        MKQL_ENSURE(joinKind == EJoinKind::Inner || joinKind == EJoinKind::Left || joinKind == EJoinKind::Right || joinKind == EJoinKind::Full, "Unsupported join kind");
         Pointers_.resize(LeftColumnTypes_.size());
         for (int index = 0; index < std::ssize(LeftKeyColumns_); ++index) {
             Pointers_[LeftKeyColumns_[index]] = &Values_[index];
         }
         int valuesIndex = 0;
-        for(int index = 0; index < std::ssize(Pointers_); ++index){
-            if (!Pointers_[index]){
+        for(int index = 0; index < std::ssize(Pointers_); ++index) {
+            if (!Pointers_[index]) {
                 Pointers_[index] = &Values_[ std::ssize(LeftKeyColumns_) + valuesIndex];
                 valuesIndex++;
             }
@@ -118,22 +148,32 @@ public:
             switch (result) {
             case EFetchResult::Finish: {
                 LeftFinished_ = true;
-                return EFetchResult::Finish;
+                if (Table_.UnusedTrackingOn()) {
+                    Table_.ForEachUnused([this](NJoinTable::TTuple unused) {
+                        AppendTuple(nullptr, unused, Output_);
+                    });
+                }
+                return EFetchResult::Yield;
             }
             case EFetchResult::Yield: {
                 return EFetchResult::Yield;
             }
             case EFetchResult::One: {
-                Table_.Lookup(Values_.data(), [this](NJoinTable::TTuple matched) {
+                bool found = false;
+                Table_.Lookup(Values_.data(), [this, &found](NJoinTable::TTuple matched) {
                     AppendTuple(Values_.data(),matched,Output_);
+                    found = true;
                 });
+                if (!found && NJoinTable::NeedToTrackUnusedLeftTuples(JoinKind_)) {
+                    AppendTuple(Values_.data(), nullptr, Output_);
+                }
                 return EFetchResult::Yield;
             }
             default:
                 MKQL_ENSURE(false, "unreachable");
             }
         }
-        MKQL_ENSURE(false, "unreachable");
+        return EFetchResult::Finish;
     }
 
 private:
@@ -148,6 +188,7 @@ private:
     const NUdf::TLoggerPtr Logger_;
     const NUdf::TLogComponentId LogComponent_;
     const TKeyTypes KeyTypes_;
+    const EJoinKind JoinKind_;
     bool LeftFinished_ = false;
     bool RightFinished_ = false;
     NJoinTable::TStdJoinTable Table_;
@@ -169,7 +210,8 @@ public:
         TVector<TType*>&&           leftColumnTypes,
         TVector<ui32>&&             leftKeyColumns,
         TVector<TType*>&&           rightColumnTypes,
-        TVector<ui32>&&             rightKeyColumns
+        TVector<ui32>&&             rightKeyColumns,
+        EJoinKind                   joinKind
     )
         : TBaseComputation(mutables, nullptr, EValueRepresentation::Boxed)
         , LeftFlow_(leftFlow)
@@ -179,6 +221,7 @@ public:
         , LeftKeyColumns_(std::move(leftKeyColumns))
         , RightColumnTypes_(std::move(rightColumnTypes))
         , RightKeyColumns_(std::move(rightKeyColumns))
+        , JoinKind_(joinKind)
     {}
 
     EFetchResult DoCalculate(NUdf::TUnboxedValue& state, TComputationContext& ctx, NUdf::TUnboxedValue* const* output) const {
@@ -197,7 +240,7 @@ private:
             state = ctx.HolderFactory.Create<TScalarHashJoinState>(
                 LeftFlow_, RightFlow_, LeftKeyColumns_, RightKeyColumns_,
                 LeftColumnTypes_, RightColumnTypes_,
-                ctx, logger, logComponent);
+                ctx, logger, logComponent, JoinKind_);
     }
 
     void RegisterDependencies() const final {
@@ -213,6 +256,7 @@ private:
     const TVector<ui32>     LeftKeyColumns_;
     const TVector<TType*>   RightColumnTypes_;
     const TVector<ui32>     RightKeyColumns_;
+    const EJoinKind         JoinKind_;
 };
 
 } // namespace
@@ -248,8 +292,6 @@ IComputationWideFlowNode* WrapDqScalarHashJoin(TCallable& callable, const TCompu
     const auto joinKindNode = callable.GetInput(2);
     const auto rawKind = AS_VALUE(TDataLiteral, joinKindNode)->AsValue().Get<ui32>();
     const auto joinKind = GetJoinKind(rawKind);
-    MKQL_ENSURE(joinKind == EJoinKind::Inner,
-                "Only inner join is supported in scalar hash join prototype");
 
     const auto leftKeyColumnsLiteral = callable.GetInput(3);
     const auto leftKeyColumnsTuple = AS_VALUE(TTupleLiteral, leftKeyColumnsLiteral);
@@ -285,7 +327,8 @@ IComputationWideFlowNode* WrapDqScalarHashJoin(TCallable& callable, const TCompu
         std::move(leftFlowItems),
         std::move(leftKeyColumns),
         std::move(rightFlowItems),
-        std::move(rightKeyColumns)
+        std::move(rightKeyColumns),
+        joinKind
     );
 }
 } // namespace NKikimr::NMiniKQL