+++ yt/yt/client/job_tracker_client/public.cpp
@@ -6,6 +6,7 @@ namespace NYT::NJobTrackerClient {
 
 const TJobId NullJobId;
 const TOperationId NullOperationId;
+const TJobTraceId NullJobTraceId;
 
 ////////////////////////////////////////////////////////////////////////////////